22.12 update

+++ ELibrary/Models/Books.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Configuration;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Web;
 
@@ -10,11 +12,11 @@ namespace ELibrary.Models
         public int ISBN { get; set; }
         public string Title { get; set; }
         public string Authors { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public int PriceDecrease { get; set; }
         public string Cover { get; set; }
         public string Publisher { get; set; }
-        public DateTime PublishYear {  get; set; }
+        public DateTime PublishYear { get; set; }
         public string Genre { get; set; }
         public bool IsBuyOnly { get; set; }
         public string Desrip { get; set; }
